Fix slider behavior for color items. (#1660)
* Fix slider behavior for color items.

The item update sent e.g. '45.0' as brightness, which then yielded a HSB state string
like '0,0,45.0', which we couldn't parse when getting it back. Fix the command sending
to send the (integer) progress instead, as this matches what we're doing when actually
binding the view.

Fixes #1658

* Make brightness parsing more lenient.

The HSB value pattern explicitly allows floats, so don't parse the
string to int (which throws on float values), but instead parse to
float and round that to integer.

Signed-off-by: Danny Baumann <dannybaumann@web.de>