Refactor server reachability check logic.
Instead of relying on the connection's socket factory for doing the
reachability check, pass the network to be used for the check directly
to the check method, which makes sure we aren't disturbing any ongoing
foreground operations on this connection while doing the check.

While at it, also move the checking code to DefaultConnection (it's only
needed for actual server connections, after all) and make the check
cancelable by using delay() instead of Thread.sleep().

Closes #1804

Signed-off-by: Danny Baumann <dannybaumann@web.de>