Replace item prefix with device identifier
Rename the existing prefix of "Send device info" to "Device identifier"
and make it available for NFC tags and voice command.

Closes #1692
Closes #1951

For NFC tags a new query parameter is added (`d=true`) if the device
identifier should be send instead of the state on the tag. Older clients
will send "UNSUPPORTED" instead.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>