Don't forcibly fetch images when resuming activities.
PR #1717 intended to force image load when bringing the app from
background to foreground, which unfortunately had the side effect of
fetching the image twice when first loading the page/activity (once when
attaching the view to the window, second on resume). Avoid that by
tweaking the logic a bit: reset 'last refresh' timestamp on pause of
activity, which leads to 'next refresh due time' being in the past on
next resume.

Signed-off-by: Danny Baumann <dannybaumann@web.de>