Always avoid cache when loading images that have refresh set.
We already avoided the cache for the actual refresh, but not for the
initial load triggered via onAttachedToWindow(). Since we need the
information whether refresh is enabled in that method, but the actual
sequence of calls is setImageUrl() -> onAttachedToWindow() ->
startRefreshing(), move the refresh interval setter to setImageUrl(), so
we have the interval available in onAttachedToWindow().

Closes #2066

Signed-off-by: Danny Baumann <dannybaumann@web.de>