Fix images sometimes not being loaded or refreshed.
Since we need the width of the ImageView for determining the desired
bitmap size, we defer loading the image until after the layout pass (so
the width is known). When this deferral happens, we rely on the
ImageView getting freshly laid out after loading the bitmap. However, in
the case of the image request being fulfillable from the cache, it
happened that the framework's requestLayout() call became a no-op,
because it was triggered from within a layout pass:

   at org.openhab.habdroid.ui.widget.WidgetImageView.requestLayout(WidgetImageView.kt:179)
   at android.widget.ImageView.setImageDrawable(ImageView.java:571)
   at androidx.appcompat.widget.AppCompatImageView.setImageDrawable(AppCompatImageView.java:104)
   at org.openhab.habdroid.ui.widget.WidgetImageView.setImageDrawable(WidgetImageView.kt:116)
   at android.widget.ImageView.setImageBitmap(ImageView.java:705)
   at androidx.appcompat.widget.AppCompatImageView.setImageBitmap(AppCompatImageView.java:112)
   at org.openhab.habdroid.ui.widget.WidgetImageView.setBitmapInternal(WidgetImageView.kt:216)
   at org.openhab.habdroid.ui.widget.WidgetImageView.doLoad(WidgetImageView.kt:199)
   at org.openhab.habdroid.ui.widget.WidgetImageView.onLayout(WidgetImageView.kt:103)

To avoid that from happening, defer the load operation to happen outside
the layout pass.

Closes #2066

Signed-off-by: Danny Baumann <dannybaumann@web.de>