Reduce retry count for important work (#2377)
"Important" work is usually triggered by an action of the user.
When I tap on a quick tile, e.g. `TOGGLE` light, and the connection cannot be established, I use the wall switch to turn it on. Then the 7th or so retry succeeds and `TOGGLE`s the light of again.

With this PR the error is shown faster to avoid such behavior.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>