Add theming support to widgets
It's possible to switch between a light and a dark theme. "Auto"
wouldn't work here, because widgets don't get a callback when the device
theme changes.
The last set theme is the default for new themes.

AFAIK widget layouts cannot be changed dynamically, so I had to add a
bunch of layout files.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>