Hide client cert pref by default
IMO there are more users that are confused by this settings, then
actual users of this. I added an entry "Advanced" that shows the
client cert and wifi prefs when clicked.

The Android settings have the same feature: Under "Network and Internet"
a few preferences are hidden by default, e.g. VPN or DoH.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>