Remove state from Item label (#3146)
For #3143 I set the label of an Item to `Color [JS(hsb-to-brightness.js):%s]`. When this Item is fetched via /rest/items/ the brackets aren't resolved to the actual value. When fetched via /rest/sitemaps/ the brackets are placed.
Since `[JS(hsb-to-brightness.js):%s]` isn't useful to be shown in the
device control, I added a splitting of the label as it's done for
Widgets.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>