Fix applying first map location
If the map isn't loaded yet when the first map widget is bound, the bind
attempt for that widget was discarded. Fix this by making sure to apply
the widget bind attempt after initially loading the map.
While at it, also make sure not to load the map until after the data
saver check happened to not cause data transfer before that check.

Fixes #3412

Signed-off-by: Danny Baumann <dannybaumann@web.de>