Custom targets for device control
The UI command item in Main UI uses the same URL for `navigate:` than
Main UI itself.
Therefore use the same URLs as the browser uses. When a Basic UI URL is
detected, open it in Sitemap, otherwise in Main UI Webview.

If no `link_to_more` is set, default to a list of members for groups.

Closes #3140
Closes #3529

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>