Don't truncate widget values (#3795)
The widget value is capped to 3 lines, which might not be enough space for some Text items with longer text, e.g. weather warnings (https://www.openhab.org/addons/bindings/dwdunwetter/). I don't see a reason why the value should be truncated.

This only applies to the default layout. For the compact layout, it's still limited to one line.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>