Fix some crashes in chart activity (#3898)
* Fix off-by-one error in wiget period parsing

E.g. '12h' was converted to 'PT12hH', not 'PT12H'.

Fixes #3897

* Properly initialize chart period

If widget period can not be parsed, use a default duration instead of
crashing.

* Fix potential TransactionTooLargeException for large charts

With large numbers of data points storing the chart data in the parcel
became too large. For a specific chart with ~225000 data points the
parcel data amounted to ~96 MB. Improve this by 2 measures:
- The chart needs a back-reference from a data point to its series.
  Previously this was done by an intermediate class for the series data,
  but that led to the series data being parceled for each data point.
  Avoid this by constructing the intermediate class when configuring the
  chart and not parceling it. Doing so reduces the parcel data size from
  ~96 MB to ~21 MB.
- Compress the chart data in the parcel. Achievable compression seems to
  be about 1:10, so the final size of the example chart data in the
  parcel becomes less than 2 MB.

Signed-off-by: Danny Baumann <dannybaumann@web.de>