Fix icons sometimes not being loaded correctly (#3906)
When scrolling through a widget list, an icon WidgetImageView sometimes
goes through a lifecycle like this
1 onAttachedToWindow
2 onLayout
3 onDetachedFromWindow
4 onAttachedToWindow
5 onLayout

If step 2 saw a pending load job and launched a coroutine to execute it,
it could happen that step 3 canceled the coroutine scope before the
coroutine was executed. Make sure not to forget about the pending
request, so it's reexecuted in step 5.

Signed-off-by: Danny Baumann <dannybaumann@web.de>